refactor/EV-804/foi criado uma pasta para riverpod
